store-ghost-post-id/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.17.2
